# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hostel Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const PaymentStatus = {
  pending: "pending",
  paid: "paid",
  overdue: "overdue",
} as const;

export interface Payment {
  id: number;
  studentId: number;
  studentName: string;
  amount: number;
  dueDate: string;
  paidDate?: string | null;
  status: PaymentStatus;
  description?: string | null;
  month?: string | null;
  createdAt: string;
}

export type StudentStatus = (typeof StudentStatus)[keyof typeof StudentStatus];

export const StudentStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Student {
  id: number;
  name: string;
  email: string;
  phone: string;
  address: string;
  roomId?: number | null;
  roomNumber?: string | null;
  joinDate: string;
  status: StudentStatus;
  emergencyContact?: string | null;
  emergencyPhone?: string | null;
  education?: string | null;
  studyYear?: string | null;
  finalInstallmentDate?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalStudents: number;
  activeStudents: number;
  totalRooms: number;
  availableRooms: number;
  occupiedRooms: number;
  totalPaymentsCollected: number;
  pendingPayments: number;
  overduePayments: number;
  overdueAmount: number;
  recentPayments: Payment[];
  recentStudents: Student[];
}

export interface CreateStudentBody {
  name: string;
  email: string;
  phone: string;
  address: string;
  roomId?: number | null;
  joinDate: string;
  emergencyContact?: string | null;
  emergencyPhone?: string | null;
  education?: string | null;
  studyYear?: string | null;
  finalInstallmentDate?: string | null;
}

export type UpdateStudentBodyStatus =
  (typeof UpdateStudentBodyStatus)[keyof typeof UpdateStudentBodyStatus];

export const UpdateStudentBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface UpdateStudentBody {
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  roomId?: number | null;
  joinDate?: string;
  status?: UpdateStudentBodyStatus;
  emergencyContact?: string | null;
  emergencyPhone?: string | null;
  education?: string | null;
  studyYear?: string | null;
  finalInstallmentDate?: string | null;
}

export type RoomType = (typeof RoomType)[keyof typeof RoomType];

export const RoomType = {
  single: "single",
  double: "double",
  triple: "triple",
  dormitory: "dormitory",
} as const;

export type RoomStatus = (typeof RoomStatus)[keyof typeof RoomStatus];

export const RoomStatus = {
  available: "available",
  full: "full",
  maintenance: "maintenance",
} as const;

export interface Room {
  id: number;
  number: string;
  type: RoomType;
  floor: number;
  capacity: number;
  occupied: number;
  monthlyRent: number;
  status: RoomStatus;
  amenities?: string | null;
  createdAt: string;
}

export type RoomWithStudents = Room & {
  students: Student[];
};

export type CreateRoomBodyType =
  (typeof CreateRoomBodyType)[keyof typeof CreateRoomBodyType];

export const CreateRoomBodyType = {
  single: "single",
  double: "double",
  triple: "triple",
  dormitory: "dormitory",
} as const;

export interface CreateRoomBody {
  number: string;
  type: CreateRoomBodyType;
  floor: number;
  capacity: number;
  monthlyRent: number;
  amenities?: string | null;
}

export type UpdateRoomBodyType =
  (typeof UpdateRoomBodyType)[keyof typeof UpdateRoomBodyType];

export const UpdateRoomBodyType = {
  single: "single",
  double: "double",
  triple: "triple",
  dormitory: "dormitory",
} as const;

export type UpdateRoomBodyStatus =
  (typeof UpdateRoomBodyStatus)[keyof typeof UpdateRoomBodyStatus];

export const UpdateRoomBodyStatus = {
  available: "available",
  full: "full",
  maintenance: "maintenance",
} as const;

export interface UpdateRoomBody {
  number?: string;
  type?: UpdateRoomBodyType;
  floor?: number;
  capacity?: number;
  monthlyRent?: number;
  status?: UpdateRoomBodyStatus;
  amenities?: string | null;
}

export interface CreatePaymentBody {
  studentId: number;
  amount: number;
  dueDate: string;
  description?: string | null;
  month?: string | null;
}

export type UpdatePaymentBodyStatus =
  (typeof UpdatePaymentBodyStatus)[keyof typeof UpdatePaymentBodyStatus];

export const UpdatePaymentBodyStatus = {
  pending: "pending",
  paid: "paid",
  overdue: "overdue",
} as const;

export interface UpdatePaymentBody {
  amount?: number;
  dueDate?: string;
  paidDate?: string | null;
  status?: UpdatePaymentBodyStatus;
  description?: string | null;
  month?: string | null;
}

export type ListStudentsParams = {
  roomId?: number;
  status?: ListStudentsStatus;
  search?: string;
};

export type ListStudentsStatus =
  (typeof ListStudentsStatus)[keyof typeof ListStudentsStatus];

export const ListStudentsStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export type ListRoomsParams = {
  status?: ListRoomsStatus;
};

export type ListRoomsStatus =
  (typeof ListRoomsStatus)[keyof typeof ListRoomsStatus];

export const ListRoomsStatus = {
  available: "available",
  full: "full",
  maintenance: "maintenance",
} as const;

export type ListPaymentsParams = {
  studentId?: number;
  status?: ListPaymentsStatus;
};

export type ListPaymentsStatus =
  (typeof ListPaymentsStatus)[keyof typeof ListPaymentsStatus];

export const ListPaymentsStatus = {
  pending: "pending",
  paid: "paid",
  overdue: "overdue",
} as const;