# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hostel Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: roomId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Add a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Student:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        roomId:
          type: integer
          nullable: true
        roomNumber:
          type: string
          nullable: true
        joinDate:
          type: string
        status:
          type: string
          enum: [active, inactive]
        emergencyContact:
          type: string
          nullable: true
        emergencyPhone:
          type: string
          nullable: true
        education:
          type: string
          nullable: true
        studyYear:
          type: string
          nullable: true
        finalInstallmentDate:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - phone
        - address
        - joinDate
        - status
        - createdAt
    CreateStudentBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        roomId:
          type: integer
          nullable: true
        joinDate:
          type: string
        emergencyContact:
          type: string
          nullable: true
        emergencyPhone:
          type: string
          nullable: true
        education:
          type: string
          nullable: true
        studyYear:
          type: string
          nullable: true
        finalInstallmentDate:
          type: string
          nullable: true
      required:
        - name
        - email
        - phone
        - address
        - joinDate
    UpdateStudentBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        roomId:
          type: integer
          nullable: true
        joinDate:
          type: string
        status:
          type: string
          enum: [active, inactive]
        emergencyContact:
          type: string
          nullable: true
        emergencyPhone:
          type: string
          nullable: true
        education:
          type: string
          nullable: true
        studyYear:
          type: string
          nullable: true
        finalInstallmentDate:
          type: string
          nullable: true
...