# lib/db/src/schema/students.ts
import { pgTable, serial, text, integer, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { roomsTable } from "./rooms";

export const studentsTable = pgTable("students", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  email: text("email").notNull().unique(),
  phone: text("phone").notNull(),
  address: text("address").notNull(),
  roomId: integer("room_id").references(() => roomsTable.id, { onDelete: "set null" }),
  joinDate: text("join_date").notNull(),
  status: text("status").notNull().default("active"),
  emergencyContact: text("emergency_contact"),
  emergencyPhone: text("emergency_phone"),
  education: text("education"),
  studyYear: text("study_year"),
  finalInstallmentDate: text("final_installment_date"),
  createdAt: timestamp("created_at").notNull().defaultNow(),
});

export const insertStudentSchema = createInsertSchema(studentsTable).omit({ id: true, createdAt: true });
export type InsertStudent = z.infer<typeof insertStudentSchema>;
export type Student = typeof studentsTable.$inferSelect;